Cleanup and unify start, build and generate book scripts
